Move tests to new layout and refactor
