+++ control_panel_api/tests/test_permissions.py
@@ -6,25 +6,22 @@ from rest_framework.status import (
     HTTP_200_OK,
     HTTP_201_CREATED,
     HTTP_204_NO_CONTENT,
-    HTTP_400_BAD_REQUEST,
     HTTP_403_FORBIDDEN,
 )
 from rest_framework.test import APITestCase
 
-from control_panel_api.aws import aws
-from control_panel_api.models import AccessToS3Bucket, AppS3Bucket
+from control_panel_api.models import AppS3Bucket
 
 
-@patch.object(aws, 'client', MagicMock())
+@patch('control_panel_api.aws.aws.client', MagicMock())
 class AppPermissionsTest(APITestCase):
     def setUp(self):
         super().setUp()
-        # Create users
         self.superuser = mommy.make(
             'control_panel_api.User', is_superuser=True)
         self.normal_user = mommy.make(
             'control_panel_api.User', is_superuser=False)
-        # Create some apps
+
         self.app_1 = mommy.make(
             "control_panel_api.App", name="App 1")
         self.app_2 = mommy.make(
@@ -54,8 +51,7 @@ class AppPermissionsTest(APITestCase):
         response = self.client.get(reverse('app-detail', (self.app_1.id,)))
         self.assertEqual(HTTP_403_FORBIDDEN, response.status_code)
 
-    @patch('boto3.client')
-    def test_delete_as_superuser_responds_OK(self, mock_client):
+    def test_delete_as_superuser_responds_OK(self):
         self.client.force_login(self.superuser)
 
         response = self.client.delete(
@@ -69,8 +65,7 @@ class AppPermissionsTest(APITestCase):
             reverse('app-detail', (self.app_1.id,)))
         self.assertEqual(HTTP_403_FORBIDDEN, response.status_code)
 
-    @patch('boto3.client')
-    def test_create_as_superuser_responds_OK(self, mock_client):
+    def test_create_as_superuser_responds_OK(self):
         self.client.force_login(self.superuser)
 
         data = {'name': 'foo', 'repo_url': 'https://example.com'}
@@ -101,28 +96,27 @@ class AppPermissionsTest(APITestCase):
         self.assertEqual(HTTP_403_FORBIDDEN, response.status_code)
 
 
-@patch.object(aws, 'client', MagicMock())
+@patch('control_panel_api.aws.aws.client', MagicMock())
 class AppS3BucketPermissionsTest(APITestCase):
     def setUp(self):
         super().setUp()
-        # Create users
         self.superuser = mommy.make(
             "control_panel_api.User", is_superuser=True)
         self.normal_user = mommy.make(
             "control_panel_api.User", is_superuser=False)
-        # Create some apps
+
         self.app_1 = mommy.make(
             "control_panel_api.App", name="App 1")
         self.app_2 = mommy.make(
             "control_panel_api.App", name="App 2")
-        # Create some S3 buckets
+
         self.s3bucket_1 = mommy.make(
             "control_panel_api.S3Bucket", name="test-bucket-1")
         self.s3bucket_2 = mommy.make(
             "control_panel_api.S3Bucket", name="test-bucket-2")
         self.s3bucket_3 = mommy.make(
             "control_panel_api.S3Bucket", name="test-bucket-3")
-        # Grant access to these S3 buckets
+
         self.apps3bucket_1 = self.app_1.apps3buckets.create(
             s3bucket=self.s3bucket_1,
             access_level=AppS3Bucket.READONLY,
@@ -208,158 +202,6 @@ class AppS3BucketPermissionsTest(APITestCase):
         data = {'doesnt': 'matter'}
         response = self.client.put(
             reverse('apps3bucket-detail', (self.apps3bucket_1.id,)), data)
-        self.assertEqual(HTTP_403_FORBIDDEN, response.status_code)
-
-
-@patch('control_panel_api.aws.aws.client', MagicMock())
-class UserS3Buckets(APITestCase):
-    def setUp(self):
-        super().setUp()
-        self.superuser = mommy.make(
-            'control_panel_api.User',
-            auth0_id='github|user_1',
-            is_superuser=True,
-        )
-        self.normal_user = mommy.make(
-            'control_panel_api.User',
-            username='alice normal',
-            auth0_id='github|user_2',
-            is_superuser=False,
-        )
-
-        self.user_1 = mommy.make(
-            "control_panel_api.User",
-            is_superuser=False,
-        )
-        self.s3bucket_1 = mommy.make(
-            "control_panel_api.S3Bucket", name="test-bucket-1")
-        self.s3bucket_2 = mommy.make(
-            "control_panel_api.S3Bucket", name="test-bucket-2")
-
-        self.users3bucket_admin = mommy.make(
-            "control_panel_api.UserS3Bucket",
-            user=self.normal_user,
-            s3bucket=self.s3bucket_1,
-            access_level=AccessToS3Bucket.READWRITE,
-            is_admin=True,
-        )
-        self.users3bucket_non_admin = mommy.make(
-            "control_panel_api.UserS3Bucket",
-            user=self.normal_user,
-            s3bucket=self.s3bucket_2,
-            access_level=AccessToS3Bucket.READWRITE,
-            is_admin=False,
-        )
-        self.users3bucket_2_admin = mommy.make(
-            "control_panel_api.UserS3Bucket",
-            user=self.user_1,
-            s3bucket=self.s3bucket_1,
-            access_level=AccessToS3Bucket.READWRITE,
-            is_admin=True,
-        )
-
-    def test_create_superuser_ok(self):
-        self.client.force_login(self.superuser)
-
-        data = {
-            'user': self.user_1.auth0_id,
-            's3bucket': self.s3bucket_2.id,
-            'access_level': AccessToS3Bucket.READWRITE,
-        }
-        response = self.client.post(reverse('users3bucket-list'), data)
-        self.assertEqual(HTTP_201_CREATED, response.status_code)
-
-    def test_create_normal_user_bad_data_400(self):
-        self.client.force_login(self.normal_user)
-
-        data = {'doesnt': 'matter'}
-        response = self.client.post(reverse('users3bucket-list'), data)
-        self.assertEqual(HTTP_400_BAD_REQUEST, response.status_code)
-
-    def test_create_normal_user_other_admin_ok(self):
-        self.client.force_login(self.normal_user)
-
-        data = {
-            'user': mommy.make('control_panel_api.User').auth0_id,
-            's3bucket': self.s3bucket_1.id,
-            'access_level': AccessToS3Bucket.READWRITE,
-            'is_admin': True,
-        }
-        response = self.client.post(reverse('users3bucket-list'), data)
-        self.assertEqual(HTTP_201_CREATED, response.status_code)
-
-    def test_create_normal_user_non_admin_403(self):
-        self.client.force_login(self.normal_user)
-
-        data = {
-            'user': mommy.make('control_panel_api.User').auth0_id,
-            's3bucket': self.s3bucket_2.id,
-            'access_level': AccessToS3Bucket.READWRITE,
-            'is_admin': True,
-        }
-        response = self.client.post(reverse('users3bucket-list'), data)
-        self.assertEqual(HTTP_403_FORBIDDEN, response.status_code)
-
-    def test_delete_super_user_owner_admin_ok(self):
-        self.client.force_login(self.superuser)
-
-        response = self.client.delete(
-            reverse('users3bucket-detail', (self.users3bucket_admin.id,)))
-        self.assertEqual(HTTP_204_NO_CONTENT, response.status_code)
-
-    def test_delete_normal_user_owner_admin_ok(self):
-        self.client.force_login(self.normal_user)
-
-        response = self.client.delete(
-            reverse('users3bucket-detail', (self.users3bucket_admin.id,)))
-        self.assertEqual(HTTP_204_NO_CONTENT, response.status_code)
-
-    def test_delete_normal_user_other_admin_ok(self):
-        self.client.force_login(self.normal_user)
-
-        response = self.client.delete(
-            reverse('users3bucket-detail', (self.users3bucket_2_admin.id,)))
-        self.assertEqual(HTTP_204_NO_CONTENT, response.status_code)
-
-    def test_delete_normal_user_owner_non_admin_403(self):
-        self.client.force_login(self.normal_user)
-
-        response = self.client.delete(
-            reverse('users3bucket-detail', (self.users3bucket_non_admin.id,)))
-        self.assertEqual(HTTP_403_FORBIDDEN, response.status_code)
-
-    def test_update_super_user_owner_admin_ok(self):
-        self.client.force_login(self.superuser)
-
-        data = {'access_level': AccessToS3Bucket.READONLY}
-        response = self.client.patch(
-            reverse('users3bucket-detail', (self.users3bucket_admin.id,)), data)
-        self.assertEqual(HTTP_200_OK, response.status_code)
-
-    def test_update_normal_user_owner_admin_ok(self):
-        self.client.force_login(self.normal_user)
-
-        data = {'access_level': AccessToS3Bucket.READONLY}
-        response = self.client.patch(
-            reverse('users3bucket-detail', (self.users3bucket_admin.id,)), data)
-        self.assertEqual(HTTP_200_OK, response.status_code)
-
-    def test_update_normal_user_other_admin_ok(self):
-        self.client.force_login(self.normal_user)
-
-        data = {'access_level': AccessToS3Bucket.READONLY}
-        response = self.client.patch(
-            reverse('users3bucket-detail', (self.users3bucket_2_admin.id,)),
-            data)
-        self.assertEqual(HTTP_200_OK, response.status_code)
-
-    def test_update_normal_user_non_admin_403(self):
-        self.client.force_login(self.normal_user)
-
-        data = {'access_level': AccessToS3Bucket.READONLY}
-        response = self.client.patch(
-            reverse('users3bucket-detail', (self.users3bucket_non_admin.id,)),
-            data)
         self.assertEqual(HTTP_403_FORBIDDEN, response.status_code)
